Divided the two R files
One is now "positive regression" and the other is now "negative
regression". All ANOVAs are also now beta regression, as the y is a 0,1
value. This allows access to a T value, as well as a P value (even
though most of the time the P value is the same as the Linear Model P
value). SS was kept for comparison's sake.